refactor: extract virtual-list core CSS to styles folder
