+++ packages/virtual-list/src/styles/vaadin-virtual-list-styles.js
